Use slightly different options when translating to prevent i18next from trying to load languages before init() is called when options.lng is changed.
At least when using a custom loader, when options.lng is changed, Angular's digest cycle will try to translate the keys in the page before the new language has been loaded by init() (called from the watcher on options). It results on i18next trying to load the new language _for every single translation on the page_, which can cause thousands of call on the custom loading function. Not passing the language when translating will make i18next use the last one specified, and the new language will be loaded when the digest loop hits the language change in the options, causing a new init(). This way the custom loading function is only called once for every new language/namespace, and the keys will be retranslated when the new language is set (in init()'s callback).